+++ src/main/java/com/tjoeun/util/PaginationUtil.java
@@ -5,18 +5,18 @@ import org.springframework.ui.Model;
 
 public class PaginationUtil {
 
-    public static void setPaging(Model model, Page<?> page, String url, String customSort, int visiblePageCount) {
+    public static void setPaging(Model model, Page<?> page, String url) {
         int currentPage = page.getNumber();         // 0-based 현재 페이지
         int totalPages = page.getTotalPages();      // 전체 페이지 수
 
-        int startPage = Math.max(0, currentPage - visiblePageCount / 2);
-        int endPage = Math.min(totalPages - 1, startPage + visiblePageCount - 1);
-        startPage = Math.max(0, endPage - visiblePageCount + 1); // 보정
+        // 페이지 번호 범위 (10개 단위)
+        int startPage = (currentPage / 10) * 10;
+        int endPage = Math.min(startPage + 9, totalPages - 1);
 
+        // model에 속성 등록
+        model.addAttribute("jobPage", page);
         model.addAttribute("startPage", startPage);
         model.addAttribute("endPage", endPage);
-        model.addAttribute("url", url);
-        model.addAttribute("customSort", customSort);
+        model.addAttribute("url", url);  // 페이징 링크에 사용할 base URL
     }
-
 }
