+++ src/main/resources/templates/empl/empl_main.html
@@ -86,7 +86,7 @@
 <!--       </div>-->
 <!--     </div>-->
 
-    <div th:replace="fragments/pagination :: fragment(${jobPage},
+    <div th:replace="fragments/pagination2 :: fragment(${jobPage},
              ${startPage}, ${endPage}, '/empl/empl_main', ${customSort})">
     </div>
