+++ src/main/resources/templates/mypage/apply_status.html
@@ -47,7 +47,7 @@
 
       <!-- 페이지네이션  -->
       <div th:replace="~{fragments/pagination :: fragment(
-           ${jobPage}, ${startPage}, ${endPage}, '/mypage/apply_status', ${customSort})}">
+           ${jobPage}, ${startPage}, ${endPage}, '/mypage/apply_status')}">
       </div>
     </div>
   </div>
