+++ src/main/resources/templates/mypage/scrap.html
@@ -95,7 +95,7 @@
         </div>
 
         <!-- 페이지네이션 -->
-        <div th:replace="~{fragments/pagination :: fragment(${jobPage}, ${startPage}, ${endPage}, '/mypage/scrap', ${customSort})}"></div>
+        <div th:replace="~{fragments/pagination :: fragment(${jobPage}, ${startPage}, ${endPage}, '/mypage/scrap')}"></div>
 
       </div>
     </div>
